Add support for highligher queries
MonitorQuery objects can be created with an optional highlighter query, that
will be run over the input document if there is a match in the original query.
If the highlighter query produces no matches, then the matches already retrieved
by the original query will be returned; otherwise the highlighter matches will
be returned.