Major refactoring of the API
You now pass in a CandidateMatcher to match(), which does it's own reporting of
how the various candidate queries match a document.

The Monitor takes a query parser as a dependency, and uses that to parse text
strings on update.  Parsed queries are stored in a Cache, keyed by the textual
form of the query, which is stored in the query index as a docvalues field.