Add the ability to give certain fields lower weight when selecting terms to index
Say you have a query of the form:
  category:a AND text:foo*

Because text:foo* is a wildcard term, category:a is indexed in preference.
However, this will end up being run on every document in category:a, which may
be most of them.  This commit allows you to pass a set of fields to the
BooleanTermExtractor which should be weighted lower when selecting which terms
to index.