Allow the Monitor to purge unused queries
This is done by running a query against the monitor's queryindex and collecting
all queries referenced, and then swapping this collection out for the previous
cache.

To be threadsafe, updates to the queryindex are now done while holding the
readlock of a ReadWriteLock, and the cache holds the writelock firstly during
the creation of an update log, and then when the update log is copied to the
new cache and the caches are swapped.