Add further shortcut to SuffixingNGramTokenFilter
We add a second cache to the TokenFilter, this time for all seen grams.  If a
gram is encountered that is already in this cache, then move directly to the
next gram starting point.

This has the effect of removing all duplicates in-place, so there is no longer a
need for the DuplicateRemovingTokenFilter (with the exception of wildcard
tokens, which should not be frequent enough to cause problems).