Add immutable Matches<T> object to hold responses
Rather than returning the CandidateMatcher, which has a bunch of methods on it
only of interest to the Monitor itself, we return an immutable Matches<T> object
from Monitor.match(InputDocument,MatcherFactory).

This also allows us to parametrize everything by QueryMatch type, rather than
having to add in CandidateMatcher types as well, which simplifies factory
creation and delegation.