Improve disjunction build times using a per-reader QueryTermFilter
Disjunction build times were being dominated by TermsEnum seekExact calls.  To
avoid this, we now build a hash of terms per field for the query index, and pass
that down to the Presearcher.  Now only terms that we know exist are used in the
disjunction, which yields an approximate x2 speedup in benchmarking.

Also cuts over TermFilteredPresearcher to use a TermsQuery